+++ packages/orm/angel_orm_mysql/test/models/has_map.g.dart
@@ -157,7 +157,7 @@ class HasMapQueryValues extends MapQueryValues {
 // **************************************************************************
 
 @generatedSerializable
-class HasMap implements _HasMap {
+class HasMap implements HasMapEntity {
   HasMap({
     this.value,
     this.list = const [],
@@ -178,7 +178,7 @@ class HasMap implements _HasMap {
 
   @override
   bool operator ==(other) {
-    return other is _HasMap &&
+    return other is HasMapEntity &&
         MapEquality<dynamic, dynamic>(
                 keys: DefaultEquality(), values: DefaultEquality())
             .equals(other.value, value) &&
@@ -242,7 +242,7 @@ class HasMapSerializer extends Codec<HasMap, Map> {
             : []);
   }
 
-  static Map<String, dynamic> toMap(_HasMap? model) {
+  static Map<String, dynamic> toMap(HasMapEntity? model) {
     if (model == null) {
       throw FormatException("Required field [model] cannot be null");
     }